Remove gslsys-mfnlin and move to gsl-sys crate.
