--- gslsys-mfnlin/Cargo.toml
@@ -1,17 +1,0 @@
-[package]
-name = "gslsys-mfnlin"
-version = "0.1.0"
-edition = "2021"
-
-build = "build.rs"
-
-[dependencies]
-libc = "0.2"
-GSL-sys = { version = "3.0.0", path = "../gsl-sys" }
-
-[build-dependencies]
-pkg-config = "0.3"
-
-[lib]
-name = "gslsys_mfnlin"
-crate-type = ["dylib", "rlib"]
